freq to midi complete tests + cmake

+++ include/frequency_to_midi/frequency_to_midi.hpp
@@ -6,7 +6,7 @@ const int min_midi = 40;
 const int max_midi = 86;
 
 struct Triple {
-    int num;
+    size_t num;
     int midi;
     double duration;
 };
